% File src/library/utils/man/Question.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2009 R Core Development Team
% Distributed under GPL 2 or later

\name{Question}
\alias{Question}
\alias{?}
\alias{print.help_files_with_topic}
\title{Documentation Shortcuts}
\description{
  These functions provide access to documentation.
  Documentation on a topic with name \code{name} (typically, an \R
  object or a data set) can be displayed by either \code{help("name")} or
  \code{?name}. For complete details, see \code{\link[utils]{Question}}.
}
% it is in fact `?`(e1, e2)
\keyword{documentation}
