\name{revoPlot}
\title{Generic X-Y Plotting}
\alias{revoPlot}
\alias{revoPlot.default}
\alias{revoPlot.data.frame}
\alias{revoPlot.matrix}
\alias{revoPlot.ts}
\description{Microsoft R Server wrapper functions for plotting from the R 
Productivity Environment. These are lightweight wrappers to the 
\code{\link[graphics]{plot}} function; they are not intended to be called 
directly by users.}
\keyword{hplot}
